requirements.txt
apache-airflow==2.7.0
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
aiohttp==3.8.5
asyncio
python-dotenv==1.0.0